Jelle

+++ app/lib/screens/members/components/user.dart
@@ -2,7 +2,7 @@
     final int id;
     final String name;
     final String nickname;
-    final int credits;
+    final double credits;
 
     User({required this.id, required this.name, required this.nickname, required this.credits});
   }
